+++ kernel/src/arch/x86_64/interrupts/IRQ.hpp
@@ -3,7 +3,7 @@
 
 #include "isr.hpp"
 
-typedef void (*x86_64_IRQHandler_t)(x86_64_Registers* regs);
+typedef void (*x86_64_IRQHandler_t)(x86_64_Interrupt_Registers* regs);
 
 void x86_64_IRQ_Initialize();
 void x86_64_IRQ_RegisterHandler(const uint8_t irq, x86_64_IRQHandler_t handler);