+++ kernel/src/arch/x86_64/interrupts/isr.hpp
@@ -5,7 +5,7 @@
 
 #include "IDT.hpp"
 
-struct x86_64_Registers {
+struct x86_64_Interrupt_Registers {
     uint64_t ds;                                         // data segment pushed by us
     uint64_t CR2;                                        // CR2 pushed by us. useful for page faults
     uint64_t R15, R14, R13, R12, R11, R10, R9, R8, RDI, RSI, RBP, RSP, RDX, RBX, RCX, RAX; // pushaq macro
@@ -13,12 +13,12 @@ struct x86_64_Registers {
     uint64_t rip, cs, rflags, rsp, ss;                   // pushed automatically by CPU
 } __attribute__((packed));
 
-typedef void (*x86_64_ISRHandler_t)(x86_64_Registers* regs);
+typedef void (*x86_64_ISRHandler_t)(x86_64_Interrupt_Registers* regs);
 
 void x86_64_ISR_Initialize();
 void x86_64_ISR_RegisterHandler(uint8_t interrupt, x86_64_ISRHandler_t handler);
 
-extern "C" void x86_64_ISR_Handler(x86_64_Registers regs);
+extern "C" void x86_64_ISR_Handler(x86_64_Interrupt_Registers regs);
 
 void x86_64_ISR_InitializeGates();